refactor: move logic to presenter component
