+++ client/src/components/MenuPresenter/MenuView.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 
 import '../../configs/firebase-config';
 import './Menu.scss';
@@ -10,85 +10,19 @@ import { faSignOutAlt } from '@fortawesome/free-solid-svg-icons';
 import MenuIconPresenter from './MenuIcons/MenuIconPresenter';
 
 import Button, { ButtonTypes } from '../../components/button/button';
+interface Props {
+	onFailure: Function;
+	onSignIn: Function;
+	onSignOut: Function;
+	signedIn: Boolean;
+}
 
-import {
-	getAuth,
-	onAuthStateChanged,
-	signInWithCredential,
-	signOut,
-	GoogleAuthProvider,
-} from 'firebase/auth';
-
-interface Props {}
-
-const MenuView: React.FC<Props> = () => {
-	const auth = getAuth();
-	const [signedIn, setSignedIn] = useState<boolean>();
-
-	onAuthStateChanged(auth, (user) => {
-		if (user) {
-		  setSignedIn(true)
-		}
-	  });
-
-	const onFailure = (error: any) => {
-		console.log(error) // TODO: better error handling
-	}
-
-	const onSignIn = (googleUser: { [k: string]: any }) => {
-			console.log('Google Auth Response', googleUser);
-			// We need to register an Observer on Firebase Auth to make sure auth is initialized.
-			const unsubscribe = onAuthStateChanged(auth, (firebaseUser) => {
-				unsubscribe();
-				// Check if we are already signed-in Firebase with the correct user.
-				if (!isUserEqual(googleUser, firebaseUser)) {
-					// Build Firebase credential with the Google ID token.
-					const credential = GoogleAuthProvider.credential(
-						googleUser.getAuthResponse().id_token
-					);
-
-					// Sign in with credential from the Google user.
-					signInWithCredential(auth, credential).catch((error) => {
-						// Handle Errors here.
-						const errorCode = error.code;
-						const errorMessage = error.message;
-						const email = error.email;
-						const credential = GoogleAuthProvider.credentialFromError(error);
-					});
-					setSignedIn(true);
-				} else {
-          setSignedIn(true);
-					console.log('User already signed-in Firebase.');
-				}
-			});
-	};
-
-	const isUserEqual = (googleUser: any, firebaseUser: any) => {
-		if (firebaseUser) {
-			const providerData = firebaseUser.providerData;
-			for (let i = 0; i < providerData.length; i++) {
-				if (
-					providerData[i].providerId === GoogleAuthProvider.PROVIDER_ID &&
-					providerData[i].uid === googleUser.getBasicProfile().getId()
-				) {
-					// We don't need to reauth the Firebase connection.
-					return true;
-				}
-			}
-		}
-		return false;
-	};
-
-	const onSignOut = () => {
-		signOut(auth)
-			.then(() => {
-				setSignedIn(false)
-			})
-			.catch((error) => {
-        console.log(error) // TODO: better error handling
-			});
-	};
-
+const MenuView: React.FC<Props> = ({
+	onFailure,
+	onSignIn,
+	onSignOut,
+	signedIn,
+}) => {
 	return (
 		<div className='Menu'>
 			<div className='Menu-content'>
@@ -111,10 +45,10 @@ const MenuView: React.FC<Props> = () => {
 					</Button>
 				) : (
 					<GoogleLogin
-						clientId= '209447824082-unncgfvgil8pbbdr7it1vd1hlhapnp18.apps.googleusercontent.com'
+						clientId='209447824082-unncgfvgil8pbbdr7it1vd1hlhapnp18.apps.googleusercontent.com'
 						buttonText='Login with Google'
-						onSuccess={onSignIn}
-						onFailure={onFailure}
+						onSuccess={(res) => onSignIn(res)}
+						onFailure={(err) => onFailure(err)}
 						cookiePolicy={'single_host_origin'}
 					/>
 				)}
